requirements.txt
# Runtime dependencies for DFARS Desktop.
# Build/dev tools (PyInstaller, etc.) live in requirements-dev.txt.

# ─── Phase 1: Desktop shell ──────────────────────────────────────────
Flask>=3.0,<4.0
Jinja2>=3.1,<4.0
pywebview>=5.0,<6.0
platformdirs>=4.2,<5.0

# ─── Phase 2: Authentication ─────────────────────────────────────────
argon2-cffi>=23.1,<24.0

# ─── Phase 3: MFA ────────────────────────────────────────────────────
pyotp>=2.9,<3.0
qrcode[pil]>=7.4,<8.0

# ─── Phase 4: Agent Zero integration ─────────────────────────────────
# DFARS uses Agent Zero (via the _dfars_integration plugin) for all LLM
# work, so we don't bundle litellm or any model SDK. We only need:
keyring>=25.0,<26.0          # Windows Credential Manager for Fernet key storage
cryptography>=42.0,<46.0     # Fernet encryption of TOTP secret + Agent Zero API key
pywin32>=306; sys_platform == "win32"  # required by keyring's Windows backend
PyYAML>=6.0,<7.0             # config files (Phase 4 plugin client also uses it)

# ─── Phase 6: Auto-updates (wired up in Phase 6) ─────────────────────
requests>=2.32,<3.0
packaging>=24.0,<26.0
